requirements.txt
# Audio Processing
pydub==0.25.1
ffmpeg-python==0.2.0
sounddevice==0.4.6

# API Integration
elevenlabs==0.2.24
spotipy==2.23.0
openai==1.3.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pydantic==1.10.13

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging & Monitoring
loguru==0.7.2 